Structure a folder
This will change the folder structure of solvedAC.
