+++ scripts/solvedAC/constants/actions.js
@@ -1,0 +1,3 @@
+export const ACTIONS = {
+  GET_PROBLEM_INFO_BY_PROBLEM_ID: 'getProblemInfoByProblemId',
+};
